pyproject.toml
[tool.poetry]
name = "space-xchange"
version = "0.1.0"
description = ""
authors = ["newbee1905 <beenewminh@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.103.1"}
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pydantic = "^2.3.0"
sqlalchemy = "^2.0.21"
mysql-connector-python = "^8.1.0"
pymysql = "^1.1.0"
argon2-cffi = "^23.1.0"
sonyflake-py = "^1.3.0"
python-dotenv = "^1.0.0"
python-paseto = "^0.5.2"
msgpack = "^1.0.7"
web3 = "^6.11.0"
py-solc-x = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
